+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## unreleased
-* Update `jackson-jr` to v2.9.9 to fix CVE-2018-11307
+## 2.106.0
+* Add `PROCESSOR_DOES_NOT_SUPPORT_MOTO_FOR_CARD_TYPE` to validation errors
+* Update `jackson-jr` to v2.9.9 to fix CVE-2018-11307 (#75 thanks @sehrope)
+* Add `threeDSecureAuthenticationId` to `ThreeDSecureInfo`
+* Add `acsTransactionId` to `ThreeDSecureInfo`
+* Add `paresStatus` to `ThreeDSecureInfo`
+* Add `threeDSecureServerTransactionId` to `ThreeDSecureInfo`
+* Add `threeDSecureLookupInfo` to `ThreeDSecureInfo`
+* Add `threeDSecureAuthenticateInfo` to `ThreeDSecureInfo`
 
 ## 2.105.0
 * Add `merchantAccountId` setter to `ThreeDSecureLookupRequest`